JANITORIAL: Add some long casts to match %ld format
Needed on Win64 compilation.

+++ audio/decoders/qdm2.cpp
@@ -2481,7 +2481,7 @@ void QDM2Stream::qdm2_synthesis_filter(uint8 index)
 }
 
 bool QDM2Stream::qdm2_decodeFrame(Common::SeekableReadStream &in, QueuingAudioStream *audioStream) {
-	debug(1, "QDM2Stream::qdm2_decodeFrame in.pos(): %ld in.size(): %ld", in.pos(), in.size());
+	debug(1, "QDM2Stream::qdm2_decodeFrame in.pos(): %ld in.size(): %ld", long(in.pos()), long(in.size()));
 	int ch, i;
 	const int frame_size = (_sFrameSize * _channels);
